Move test function to the test directory
Move the VDI automaton test to the corresponding test file from its
source file. The all_pairs and related variables are only used by the
test, but the (-) and superstate functions are used by code in xapi.

Also, simplify the all_pairs function by using maps and flatten to make
it more readable.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>